fix: Use representations during distance calculation and return correct indices (#18)
Before, the matrices itself where used in the distance calculations of the RBF kernel. In the original paper, they used features and not raw data. Therefore, I changed the implementation to use the representations instead of the raw data.

Wrong indices were returned when the clustering method was used to select the final indices.

Some performance optimizations:
* bigger batches
* sigma calculation
 * rbf distance calculation

* fix: use representations instead of matrices to calc distances

* fix: improve performance and use l2 distance instead of l1 distance

* fix: return correct indices for knn cluster method

* fix: performance improvements of sigma calculation

* fix: performance improvements of sigma calculation

* fix: use better batch size for performance improvements